联调过程中问题进行调整

+++ uadp-module-bookstore/uadp-module-bookstore-biz/src/main/java/cn/iocoder/yudao/module/bookstore/service/bookqtcodeinfo/BookQtcodeInfoServiceImpl.java
@@ -1,5 +1,6 @@
 package cn.iocoder.yudao.module.bookstore.service.bookqtcodeinfo;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.img.ImgUtil;
 import cn.hutool.extra.qrcode.QrCodeUtil;
 import cn.hutool.extra.qrcode.QrConfig;
@@ -13,15 +14,15 @@ import cn.iocoder.yudao.module.bookstore.dal.dataobject.bookqtcodeinfo.BookQtcod
 import cn.iocoder.yudao.module.bookstore.dal.dataobject.bookqtcodesource.BookQtcodeSourceDO;
 import cn.iocoder.yudao.module.bookstore.dal.mysql.bookqtcodeinfo.BookQtcodeInfoMapper;
 import cn.iocoder.yudao.module.bookstore.dal.mysql.bookqtcodesource.BookQtcodeSourceMapper;
+import cn.iocoder.yudao.module.infra.enums.BookStoreErrorCodeConstants;
+import com.mchange.lang.LongUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 
 import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
 import static cn.iocoder.yudao.module.infra.enums.BookStoreErrorCodeConstants.BOOK_QTCODE_INFO_NOT_EXISTS;
@@ -99,26 +100,48 @@ public class BookQtcodeInfoServiceImpl implements BookQtcodeInfoService {
     @Transactional(rollbackFor = Exception.class)
     @Override
     public Long saveBookQtcodeInfoAndQtcodeResource(ExtraBookQtcodeInfoSaveReqVO reqVO) {
+        //书的码,如果已存在，不允许加
+        if(StringUtils.equalsIgnoreCase("00",reqVO.getDtcodeCategory())
+                && !Objects.isNull(reqVO.getBookNo())){
+            BookQtcodeInfoDO bookQtcodeInfoDO = bookQtcodeInfoMapper.selectQtByBookNo(reqVO.getBookNo());
+            if(!Objects.isNull(bookQtcodeInfoDO)) {
+                throw exception(BookStoreErrorCodeConstants.BOOK_QTCODE_INFO_EXISTS);
+            }
+        } else if(StringUtils.equalsIgnoreCase("01",reqVO.getDtcodeCategory())
+                && !Objects.isNull(reqVO.getChapterId())){
+            //章节的码,如果已存在，不允许加
+            BookQtcodeInfoDO bookQtcodeInfoDO = bookQtcodeInfoMapper.selectQtByChapterId(reqVO.getChapterId());
+            if(!Objects.isNull(bookQtcodeInfoDO)) {
+                throw exception(BookStoreErrorCodeConstants.CHAPTER_QTCODE_INFO_EXISTS);
+            }
+        }else{
+            //其它的码不支持
+            throw exception(BookStoreErrorCodeConstants.QTCODE_INFO_NOT_SUPPORT);
+        }
         BookQtcodeInfoDO bookQtcodeInfo = BeanUtils.toBean(reqVO, BookQtcodeInfoDO.class);
         bookQtcodeInfo.setDeptId(SecurityFrameworkUtils.getLoginUserDeptId());
         // 插入
         bookQtcodeInfoMapper.insert(bookQtcodeInfo);
-        Collection<BookQtcodeSourceDO> entities = new ArrayList<>();
-        reqVO.getSimpleBookQtcodeSourceVO().forEach(vo ->{
-            BookQtcodeSourceDO bookQtcodeSource = new BookQtcodeSourceDO();
-            bookQtcodeSource.setBookNo(reqVO.getBookNo());
-            bookQtcodeSource.setChapterId(reqVO.getBookNo());
-            bookQtcodeSource.setSourceForm(vo.getSourceForm());
-            bookQtcodeSource.setSourceName(vo.getSourceName());
-            bookQtcodeSource.setApplicaScens(vo.getApplicaScens());
-            bookQtcodeSource.setDtcodeId(bookQtcodeInfo.getId());
-            bookQtcodeSource.setSourceId(vo.getSourceId());
-            bookQtcodeSource.setBookNo(reqVO.getBookNo());
-            bookQtcodeSource.setChapterId(reqVO.getChapterId());
-            bookQtcodeSource.setDeptId(SecurityFrameworkUtils.getLoginUserDeptId());
-            entities.add(bookQtcodeSource);
-        });
-        bookQtcodeSourceMapper.insertBatch(entities);
+        if (CollUtil.isNotEmpty(reqVO.getSimpleBookQtcodeSourceVO())
+                && !Objects.isNull(reqVO.getBookNo())
+                && !Objects.isNull(reqVO.getBookNo())) {
+            Collection<BookQtcodeSourceDO> entities = new ArrayList<>();
+            reqVO.getSimpleBookQtcodeSourceVO().forEach(vo ->{
+                BookQtcodeSourceDO bookQtcodeSource = new BookQtcodeSourceDO();
+                bookQtcodeSource.setBookNo(reqVO.getBookNo());
+                bookQtcodeSource.setChapterId(reqVO.getBookNo());
+                bookQtcodeSource.setSourceForm(vo.getSourceForm());
+                bookQtcodeSource.setSourceName(vo.getSourceName());
+                bookQtcodeSource.setApplicaScens(vo.getApplicaScens());
+                bookQtcodeSource.setDtcodeId(bookQtcodeInfo.getId());
+                bookQtcodeSource.setSourceId(vo.getSourceId());
+                bookQtcodeSource.setBookNo(reqVO.getBookNo());
+                bookQtcodeSource.setChapterId(reqVO.getChapterId());
+                bookQtcodeSource.setDeptId(SecurityFrameworkUtils.getLoginUserDeptId());
+                entities.add(bookQtcodeSource);
+            });
+            bookQtcodeSourceMapper.insertBatch(entities);
+        }
         return bookQtcodeInfo.getId();
     }
 
@@ -130,33 +153,13 @@ public class BookQtcodeInfoServiceImpl implements BookQtcodeInfoService {
         //2、更新二维码信息
         BookQtcodeInfoDO updateObj = BeanUtils.toBean(reqVO, BookQtcodeInfoDO.class);
         bookQtcodeInfoMapper.updateById(updateObj);
-        //先清除资源表中此二维码的所有数据
-        bookQtcodeSourceMapper.deleteByDtcodeId(reqVO.getId());
-        Collection<BookQtcodeSourceDO> entities = new ArrayList<>();
-        reqVO.getSimpleBookQtcodeSourceVO().forEach(vo ->{
-            BookQtcodeSourceDO bookQtcodeSource = new BookQtcodeSourceDO();
-            bookQtcodeSource.setBookNo(reqVO.getBookNo());
-            bookQtcodeSource.setChapterId(reqVO.getBookNo());
-            bookQtcodeSource.setSourceForm(vo.getSourceForm());
-            bookQtcodeSource.setSourceName(vo.getSourceName());
-            bookQtcodeSource.setApplicaScens(vo.getApplicaScens());
-            bookQtcodeSource.setDtcodeId(reqVO.getId());
-            bookQtcodeSource.setSourceId(vo.getSourceId());
-            bookQtcodeSource.setBookNo(reqVO.getBookNo());
-            bookQtcodeSource.setChapterId(reqVO.getChapterId());
-            bookQtcodeSource.setDeptId(SecurityFrameworkUtils.getLoginUserDeptId());
-            entities.add(bookQtcodeSource);
-        });
-        bookQtcodeSourceMapper.insertBatch(entities);
-   /*
-        //3、更新资源表
-        reqVO.getSimpleBookQtcodeSourceVO().forEach(vo ->{
-            //说明是更新
-            if(vo.getSourceId() > 0 ){
-                BookQtcodeSourceDO bookQtcodeSourceDO = BeanUtils.toBean(vo, BookQtcodeSourceDO.class);
-                bookQtcodeSourceDO.setId(vo.getSourceId());
-                bookQtcodeSourceMapper.updateById(bookQtcodeSourceDO);
-            }else{
+        if (CollUtil.isNotEmpty(reqVO.getSimpleBookQtcodeSourceVO())
+                && !Objects.isNull(reqVO.getBookNo())
+                && !Objects.isNull(reqVO.getBookNo())) {
+            //先清除资源表中此二维码的所有数据
+            bookQtcodeSourceMapper.deleteByDtcodeId(reqVO.getId());
+            Collection<BookQtcodeSourceDO> entities = new ArrayList<>();
+            reqVO.getSimpleBookQtcodeSourceVO().forEach(vo ->{
                 BookQtcodeSourceDO bookQtcodeSource = new BookQtcodeSourceDO();
                 bookQtcodeSource.setBookNo(reqVO.getBookNo());
                 bookQtcodeSource.setChapterId(reqVO.getBookNo());
@@ -168,10 +171,10 @@ public class BookQtcodeInfoServiceImpl implements BookQtcodeInfoService {
                 bookQtcodeSource.setBookNo(reqVO.getBookNo());
                 bookQtcodeSource.setChapterId(reqVO.getChapterId());
                 bookQtcodeSource.setDeptId(SecurityFrameworkUtils.getLoginUserDeptId());
-                bookQtcodeSourceMapper.insert(bookQtcodeSource);
-            }
-        });
-        */
+                entities.add(bookQtcodeSource);
+            });
+            bookQtcodeSourceMapper.insertBatch(entities);
+        }
         return reqVO.getId();
     }
 
@@ -179,5 +182,15 @@ public class BookQtcodeInfoServiceImpl implements BookQtcodeInfoService {
     @Override
     public List<BookQtcodeSourceDO> selectListByDtcodeId(Long dtcodeId){
         return bookQtcodeSourceMapper.selectListByDtcodeId(dtcodeId);
+    }
+
+    @Override
+    public BookQtcodeInfoDO getBookQtcodeInfoByBookNo(Long bookNo) {
+        return bookQtcodeInfoMapper.selectQtByBookNo(bookNo);
+    }
+
+    @Override
+    public BookQtcodeInfoDO getBookQtcodeInfoByChapterId(Long chapterId) {
+        return bookQtcodeInfoMapper.selectOne(BookQtcodeInfoDO::getChapterId, chapterId);
     }
 }
